feat: migrate restrictions.test.js to jest
